create object_mapper.go
